Add token-from flag so that secret files can be used
This change allows Kubernetes secrets or files on disk to be used
for the token. Tested locally in Go on MacOS by creating a file
and using --token-from at both ends of the tunnel.

Beware of adding any new-lines to the file, echo -n is best for
creating a new file without any additional newline text.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>