Add control-plane option
This patch means that users can split their control plane from the
data plane. The control plane is used for the tunnel and the data
plane is used to serve traffic to external users via the
exit-node's IP.

Tested with --port 80 --control-port 8080 and seeing that data
was not served on port 8080. Regression tested that when using
the same numeric value for both ports, the existing behaviour
was kept.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>