improved startup logging with version information
streamlines the startup logging of inlets to always print the welcome
message first even if there are errors later in the process. reporting
the version number and component started helps the operator to exactly
know what's going on. printing the welcome message like the other log
lines unifies the output in a consistent way.

Signed-off-by: Tobias Brunner <tobias.brunner@vshn.ch>